requirements.txt
streamlit
python-docx
edge-tts